fix/warnings
- double -> float 수정
- 전방 선언 class, struct 구분
- 모든 경로에서 return

+++ Week-02/FRay.h
@@ -9,7 +9,7 @@ public:
 
     FRay(const FVector& _Origin, const FVector& _Direction) : Origin(_Origin), Direction(_Direction) {}
 
-    inline FVector At(double T) const
+    inline FVector At(float T) const
     {
         return Origin + T * Direction;
     }
